Grails 7 migration: fix runtime regressions found via API testing
Domains:
- Proxy-safe equals() in Project/Task/Story/Actor/Sprint/Release/Feature
  (getClass() != obj.getClass() is false against a Hibernate lazy proxy, which
  silently failed Task's parentProject/parentStory validators -> planned story
  tasks kept a null sprint / invisible on the task board)
- User.hashCode() null-safe
- uniqueResult named queries: .list() -> .get() (Story/Actor/Backlog/Feature/
  Hook/Release/Sprint/TimeBoxNotesTemplate)
- Task.getSprintBacklog() unproxies backlog; route (Sprint) task.backlog casts
- AcceptanceTest/User: coerce structural-GString HQL to String (stray :p0)

Services:
- ClicheService: Integer-key the points maps (config NavigableMap keys are
  Strings) -> fixes sprint activate null+effort NPE
- HookListenerService: Hook.async.task{} -> synchronous Hook.withSession{}
- Task/Notification/Listener services: use task.sprintBacklog
- StoryService.plan: addToTasks() so a planned story's tasks get the sprint;
  findCurrentSprint().get()
- TaskService.update: derive backlog from parentStory.parentSprint (self-heal)
- ProjectService.importXML: set preferences.project inverse
- UserService.unMarshall: merge-on-import (reuse existing account by
  uid/email/username), {SHA-256}-prefix legacy password hashes, keep IdP
  placeholders unwrapped
- AttachmentService: uniqueResult .list() -> .get()

Plugin/support:
- IcescrumCoreGrailsPlugin: guard divide-by-zero in export/import progress
- IceScrumEventPublisher: read persistentProperties from the GORM mappingContext

+++ grails-app/services/org/icescrum/core/services/ClicheService.groovy
@@ -45,8 +45,8 @@ class ClicheService {
         int allTotal = 0
         int remainingTotal = 0
         def storyTypes = grailsApplication.config.icescrum.resourceBundles.storyTypes.keySet()
-        def allPointsByType = storyTypes.collectEntries { storyType -> [(storyType): 0] }
-        def remainingPointsByType = storyTypes.collectEntries { storyType -> [(storyType): 0] }
+        def allPointsByType = storyTypes.collectEntries { storyType -> [(storyType as Integer): 0] }
+        def remainingPointsByType = storyTypes.collectEntries { storyType -> [(storyType as Integer): 0] }
         stories.each { Story story ->
             if (story.effort > 0) {
                 allPointsByType[story.type] += story.effort
@@ -119,7 +119,8 @@ class ClicheService {
                 "${Cliche.PROJECT_POINTS}"(projectPoints.allTotal)
                 // Project remaining points
                 storyTypes.each { storyType ->
-                    "${grailsApplication.config.icescrum.resourceBundles.storyTypesCliche[storyType]}"(projectPoints.remaining[storyType])
+                    // storyType comes from the config NavigableMap keySet (String in Grails 7); the points maps are Integer-keyed.
+                    "${grailsApplication.config.icescrum.resourceBundles.storyTypesCliche[storyType]}"(projectPoints.remaining[storyType as Integer])
                 }
                 "${Cliche.PROJECT_REMAINING_POINTS}"(projectPoints.remainingTotal)
                 // Release remaining points
